+++ inst/app/www/NEWS.md
@@ -1,3 +1,17 @@
+# v1.5.0
+## New features
+* Visualization of normalized data with heatmaps.
+* Option to normalize charge states of analytes separately.
+
+## Adjustments
+* Add terminal galactosylation as an automatically calculated trait for human IgG.
+* Remove the “Calculate glycosylation traits” button.
+* Disable the “Generate report” button until normalization is performed.
+
+## Bug fixes
+* Minor bug fixes.
+
+
 # v1.4.3
 ## Adjustments
 * Detect and automatically rename isomers in Skyline data.